refactor: enhance dashboard functionality
