+++ .gitignore
@@ -3,3 +3,4 @@ dist/
 coverage/
 *.tgz
 .DS_Store
+oma-dashboards/
